chore: update xui version
