+++ Jenkinsfile_CNP
@@ -117,8 +117,7 @@ withPipeline(type, product, component) {
     env.CASE_SERVICE_URL="http://fpl-case-service-pr-${CHANGE_ID}.service.core-compute-preview.internal"
     env.URL="https://xui-fpl-case-service-pr-${CHANGE_ID}.service.core-compute-preview.internal"
     env.MOCKED_PAYMENTS=false
-    env.PARALLEL_CHUNKS=5
-    env.TEST_RETRIES=2
+    env.PARALLEL_CHUNKS=11
   }
 
   after('smoketest:saat') {
