+++ Jenkinsfile_nightly
@@ -47,8 +47,7 @@ withNightlyPipeline(type, product, component) {
   before('fullFunctionalTest') {
     env.PARALLEL_CHUNKS = params.FUNCTIONAL_TESTS_WORKERS
     env.MOCKED_PAYMENTS = false
-    env.TEST_RETRIES = 5
-    env.WAIT_FOR_TIMEOUT = 60000
+    env.TEST_RETRIES = 2
     env.ENV = 'aat'
   }
